+++ app/controllers/carto/api/visualizations_controller.rb
@@ -26,14 +26,14 @@ module Carto
         :notify_watching, :list_watching, :add_like, :remove_like
 
       # TODO: compare with older, there seems to be more optional authentication endpoints
-      skip_before_filter :api_authorization_required, only: [:index, :vizjson2, :vizjson3, :is_liked, :add_like,
+      skip_before_filter :api_authorization_required, only: [:show, :index, :vizjson2, :vizjson3, :is_liked, :add_like,
                                                              :remove_like, :notify_watching, :list_watching,
                                                              :static_map]
 
       # :update and :destroy are correctly handled by permission check on the model
       before_filter :ensure_user_can_create, only: [:create]
 
-      before_filter :optional_api_authorization, only: [:index, :vizjson2, :vizjson3, :is_liked, :add_like,
+      before_filter :optional_api_authorization, only: [:show, :index, :vizjson2, :vizjson3, :is_liked, :add_like,
                                                         :remove_like, :notify_watching, :list_watching, :static_map]
 
       before_filter :id_and_schema_from_params
@@ -48,7 +48,7 @@ module Carto
                                                                          :destroy, :google_maps_static_image]
 
       before_filter :ensure_visualization_owned, only: [:destroy, :google_maps_static_image]
-      before_filter :ensure_visualization_is_viewable, only: [:add_like, :remove_like]
+      before_filter :ensure_visualization_is_likeable, only: [:add_like, :remove_like]
 
       rescue_from Carto::LoadError, with: :rescue_from_carto_error
       rescue_from Carto::UUIDParameterFormatError, with: :rescue_from_carto_error
@@ -384,7 +384,7 @@ module Carto
         raise Carto::LoadError.new('Visualization not editable', 403) unless @visualization.is_owner?(current_viewer)
       end
 
-      def ensure_visualization_is_viewable
+      def ensure_visualization_is_likeable
         return(head 403) unless current_viewer && @visualization.is_viewable_by_user?(current_viewer)
       end